Rewrite how we go about finding the old version.
Previously, we took in a .repos file to find out the "old"
version of the packages we are generating changelogs for.
But in point of fact, that doesn't make a lot of sense; that
repos file has almost certainly forked off with respect to
the rolling branch.

Instead, in this rewrite we instead ask the user to provide
us with a date that we should generate changelog since.
That date should really be the day that the previous release
(in the previous year) branched off from rolling.  That way
we ensure that there is a continuous set of commits between
today and the "since" date.  Then we look at the contributors
and changelog since that date, and output it.

This makes a small but noticeable change in the output, and
is much more correct.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>